Move FakeClock to the impl/test folder.
